+++ packages/shared/widget-plugin-component-kit/src/ThreeStateCheckBox.tsx
@@ -1,15 +1,14 @@
-import { createElement, ReactElement, useEffect, useRef } from "react";
+import { createElement, InputHTMLAttributes, ReactElement, useEffect, useRef } from "react";
+import classNames from "classnames";
 
 export type ThreeStateCheckBoxEnum = "all" | "some" | "none";
 
-export interface ThreeStateCheckBoxProps {
-    id?: string;
-    disabled?: boolean;
+export interface ThreeStateCheckBoxProps
+    extends Omit<InputHTMLAttributes<HTMLInputElement>, "checked" | "value" | "type"> {
     value: ThreeStateCheckBoxEnum;
-    onChange?: () => void;
 }
 
-export function ThreeStateCheckBox({ id, disabled, value, onChange }: ThreeStateCheckBoxProps): ReactElement {
+export function ThreeStateCheckBox({ value, className, ...props }: ThreeStateCheckBoxProps): ReactElement {
     const checkboxRef = useRef<HTMLInputElement | null>(null);
     useEffect(() => {
         if (!checkboxRef.current) {
@@ -24,13 +23,11 @@ export function ThreeStateCheckBox({ id, disabled, value, onChange }: ThreeState
 
     return (
         <input
-            id={id}
+            {...props}
             type="checkbox"
-            className="three-state-checkbox"
+            className={classNames("three-state-checkbox", className)}
             ref={checkboxRef}
             checked={value !== "none"}
-            disabled={disabled}
-            onChange={onChange}
         />
     );
 }