+++ packages/pluggableWidgets/selection-helper-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Internal improvements.
+
 ## [1.0.3] - 2023-10-13
 
 ### Fixed